requirements.txt
fastapi
uvicorn
tensorflow
tensorflow-hub
Pillow
numpy
requests
supabase
pydantic